Explicitly tell ffprobe to replace invalid UTF-8 sequences
ffprobe documentation is currently inconsistent with the source code.

According to the documentation at:
https://ffmpeg.org/ffprobe.html#Writers
the default svr / string_validation_replacement is an empty string.

But in the source code we can see that it is \xEF\xBF\xBD (the UTF-8 value
for the Unicode REPLACEMENT CHARACTER):
https://github.com/FFmpeg/FFmpeg/blob/025fcee6fa208dd484a419ecf29c3e70038a5e34/fftools/ffprobe.c#L495

Therefore, do not rely on ffprobe defaults and explicitly specify sv and svr.